style(Popup): conform more with style rfc
