+++ src/actions.js
@@ -1,13 +1,20 @@
 import 'isomorphic-fetch';
 import _ from 'underscore';
 import { requiredCerts, userapiPath, headers, basename, submissionApiOauthPath, graphqlPath } from './configs';
-import { fetchProjects, fetchDictionary } from './queryactions';
+
 
 export const updatePopup = state => ({
   type: 'UPDATE_POPUP',
   data: state,
 });
 
+export const connectionError = () => {
+  console.log('connection error');
+  return {
+    type: 'REQUEST_ERROR',
+    error: 'connection_error',
+  };
+};
 
 /**
  * Little helper issues fetch, then resolves response
@@ -20,7 +27,7 @@ export const updatePopup = state => ({
  * @return Promise<{response,data,status,headers}
  */
 export const fetchJsonOrText = (opts) => {
-  const { path, method = 'GET', body = null, customHeaders } = opts;
+  const { path, method = 'GET', body = null, customHeaders, dispatch } = opts;
 
   const request = {
     credentials: 'same-origin',
@@ -28,49 +35,52 @@ export const fetchJsonOrText = (opts) => {
     method,
     body,
   };
-  return fetch(path, request
-    ).then(
-      response => response.text().then(
-        (data) => {
-          if (data) {
-            try {
-              data = JSON.parse(data);
-            } catch (e) {
-              // # do nothing
-            }
+  return fetch(path, request,
+  ).then(
+    response => response.text().then(
+      (textData) => {
+        let data = textData;
+        if (data) {
+          try {
+            data = JSON.parse(data);
+          } catch (e) {
+            // # do nothing
           }
-          return { response, data, status: response.status, headers: response.headers };
-        }),
-        (error) => {
-          if (dispatch) { dispatch(connectionError()); }
-          return Promise.reject(error);
         }
-    );
-}
+        return { response, data, status: response.status, headers: response.headers };
+      }),
+    (error) => {
+      if (dispatch) { dispatch(connectionError()); }
+      return Promise.reject(error);
+    },
+  );
+};
 
 /**
  * Redux 'thunk' wrapper around fetchJsonOrText 
  * invokes dispatch(handler( { status, data, headers} ) and callback()
- * and propagates {response,data, status, headers} on resolved fetch, otherwise dipatch(connectionError()) on
- * fetch rejection.
+ * and propagates {response,data, status, headers} on resolved fetch, 
+ * otherwise dipatch(connectionError()) on fetch rejection.
  * May prefer this over straight call to fetchJsonOrText in Redux context due to
  * conectionError() dispatch on fetch rejection. 
  * 
  * @param {path,method=GET,body=null,customerHeaders,handler,callback} opts
  * @return Promise 
  */
-export const fetchWrapper = ({ path, method = 'GET', body = null, customHeaders, handler, callback = () => (null) }) => (dispatch) => {
-  return fetchJsonOrText({ path, method, body, customHeaders, dispatch }
-    ).then(
-      ({ response,data }) => {
-        if ( handler ) {
-          dispatch(handler({ status: response.status, data, headers: response.headers }));
-        }
-        callback();
-        return { response, data, status: response.status, headers: response.headers };
-      }
-    );
-};
+export const fetchWrapper = ({ path, method = 'GET', body = null, customHeaders, handler, callback = () => (null) }) =>
+  dispatch => fetchJsonOrText({ path, method, body, customHeaders, dispatch },
+  ).then(
+    ({ response, data }) => {
+      const result = { response, data, status: response.status, headers: response.headers };
+      const dispatchPromise = handler ? Promise.resolve(dispatch(handler(result))) : Promise.resolve('ok');
+      return dispatchPromise.then(
+        () => {
+          callback();
+          return result;
+        },
+      );
+    },
+  );
 
 export const fetchGraphQL = (graphQLParams) => {
   const request = {
@@ -109,19 +119,10 @@ export const unauthorizedError = () => ({
   error: 'unauthorized',
 });
 
-export const connectionError = () => {
-  console.log('connection error');
-  return {
-    type: 'REQUEST_ERROR',
-    error: 'connection_error',
-  };
-};
-
-
 
 export const fetchUser = dispatch => fetchJsonOrText({
   path: `${userapiPath}user/`,
-  dispatch
+  dispatch,
 }).then(
   ({ status, data }) => {
     switch (status) {
@@ -141,7 +142,7 @@ export const fetchUser = dispatch => fetchJsonOrText({
         error: data.error,
       };
     }
-  }
+  },
 ).then((msg) => { dispatch(msg); });
 
 export const requireAuth = (store, additionalHooks) => (nextState, replace, callback) => {
@@ -150,15 +151,16 @@ export const requireAuth = (store, additionalHooks) => (nextState, replace, call
     const { user } = store.getState();
     const location = nextState.location;
     if (!user.username) {
-      const path = location.pathname == '/' ? '/' : `/${location.pathname}`;
+      const path = location.pathname === '/' ? '/' : `/${location.pathname}`;
       replace({ pathname: '/login', query: { next: path + nextState.location.search } });
       return Promise.resolve();
     }
-    const has_certs = _.intersection(requiredCerts, user.certificates_uploaded).length !== requiredCerts.length;
+    const hasCerts =
+      _.intersection(requiredCerts, user.certificates_uploaded).length !== requiredCerts.length;
     // take quiz if this user doesn't have required certificate
-    if (location.pathname !== 'quiz' && has_certs) {
+    if (location.pathname !== 'quiz' && hasCerts) {
       replace({ pathname: '/quiz' });
-    } else if (location.pathname === 'quiz' && !has_certs) {
+    } else if (location.pathname === 'quiz' && !hasCerts) {
       replace({ pathname: '/' });
     } else if (additionalHooks) {
       return additionalHooks(nextState, replace);
@@ -171,17 +173,20 @@ export const requireAuth = (store, additionalHooks) => (nextState, replace, call
     .then(() => callback());
 };
 
-export const enterHook = (store, hookAction) => (nextState, replace, callback) => store.dispatch(hookAction()).then(() => callback());
+export const enterHook = (store, hookAction) =>
+  (nextState, replace, callback) => store.dispatch(hookAction()).then(() => callback());
+
+export const receiveAPILogout = handleResponse('RECEIVE_API_LOGOUT');
 
 export const logoutAPI = () => dispatch => dispatch(fetchWrapper({
   path: `${submissionApiOauthPath}logout`,
   handler: receiveAPILogout,
 })).then(() => document.location.replace(`${userapiPath}/logout?next=${basename}`));
 
-export const fetchOAuthURL = oauth_path => dispatch =>
+export const fetchOAuthURL = oauthPath => dispatch =>
 // Get cloud_middleware's authorization url
   fetchJsonOrText({
-    path: `${oauth_path}authorization_url`,
+    path: `${oauthPath}authorization_url`,
     dispatch,
   }).then(
     ({ status, data }) => {
@@ -197,7 +202,5 @@ export const fetchOAuthURL = oauth_path => dispatch =>
           error: data.error,
         };
       }
-    }
-  ).then( msg => dispatch(msg) );
-
-export const receiveAPILogout = handleResponse('RECEIVE_API_LOGOUT');
+    },
+  ).then(msg => dispatch(msg));
